Firebase config and onAuth function were moved to separate file
