+++ src/services/firebase/index.js
@@ -1,0 +1,20 @@
+import { initializeApp } from "firebase/app";
+import { getAuth, onAuthStateChanged } from "firebase/auth";
+
+const firebaseApp = initializeApp({
+  apiKey: process.env.VUE_APP_API_KEY,
+  authDomain: process.env.VUE_APP_AUTH_DOMAIN,
+  projectId: process.env.VUE_APP_PROJECT_ID,
+  storageBucket: process.env.VUE_APP_STORAGE_BUCKET,
+  messagingSenderId: process.env.VUE_APP_MESSAGING_SENDER_ID,
+  appId: process.env.VUE_APP_APP_ID,
+  measurementId: process.env.VUE_APP_MEASUREMENT_ID,
+});
+
+const auth = getAuth(firebaseApp);
+
+function onFirebaseAuthStateChanged(initFoo) {
+  onAuthStateChanged(auth, initFoo);
+}
+
+export { onFirebaseAuthStateChanged };
